Renamed TransactionIdentifier to FastIdGenerator. Moved FastIdGenerator to disco.agent.util namespace in disco-jav-agent-api package
